requirements.txt
Flask
pandas
easyocr
openpyxl
Pillow
xlsxwriter
gunicorn
